Added setup.py, modified pyproject toml to ignore tests, renamed module
