+++ pyproject.toml
@@ -17,3 +17,6 @@ profile = "black"
 [tool.bandit]
 # See rules at: https://bandit.readthedocs.io/en/1.7.4/blacklists/blacklist_calls.html
 skips = []
+
+[tool.bandit.assert_used]
+skips = ['*_test.py', '*/test_*.py']
