+++ requirements.txt
@@ -1,2 +1,3 @@
 pre-commit=3.3.3
 tensorflow=2.11.0
+pytest=7.3.1
