+++ setup.py
@@ -1,0 +1,8 @@
+from setuptools import setup, find_packages
+
+
+setup(
+    name="mlops",
+    version="1.0",
+    packages=find_packages(),
+)
