+++ test_runner.sh
@@ -1,0 +1,1 @@
+python -m pytest -v tests
